% Generated by roxygen2 (4.0.0): do not edit by hand
\name{table.Performance.pool.cran}
\alias{table.Performance.pool.cran}
\title{Print metrics from R CRAN PerformanceAnalytics that compatible with table.Performance}
\usage{
table.Performance.pool.cran(...)
}
\description{
Print metrics from R CRAN PerformanceAnalytics that
compatible with table.Performance
}
\details{
use \code{table.Performance.pool} to check available
metrics. recoded SharpeRatio
}
\examples{
table.Performance.pool.cran()
}
\author{
Kirk Li \email{kirkli@stat.washington.edu}
}
\seealso{
\code{\link{table.Performance}},
\code{\link{table.Performance.pool}}
}
\keyword{measure}
\keyword{metrics}
\keyword{performance}
\keyword{table}

