<<<<<<< HEAD
% Generated by roxygen2 (4.0.0): do not edit by hand
\name{table.Performance.pool}
\alias{table.Performance.pool}
\title{Print metrics from R-forge PerformanceAnalytics that compatible with table.Performance}
\usage{
table.Performance.pool(...)
}
\arguments{
  \item{NULL}{}
}
\description{
Print metrics from R-forge PerformanceAnalytics that
compatible with table.Performance
}
\details{
use \code{table.Performance.pool} to check available
metrics. recoded SharpeRatio
}
\examples{
table.Performance.pool()
}
\author{
Kirk Li \email{kirkli@stat.washington.edu}
}
\seealso{
\code{\link{table.Performance}},
\code{\link{table.Performance.pool.cran}}
}
\keyword{measure}
\keyword{metrics}
\keyword{performance}
\keyword{table}

=======
% Generated by roxygen2 (4.0.0): do not edit by hand
\name{table.Performance.pool}
\alias{table.Performance.pool}
\title{Print metrics from R-forge PerformanceAnalytics that compatible with table.Performance}
\usage{
table.Performance.pool(...)
}
\arguments{
  \item{NULL}{}
}
\description{
Print metrics from R-forge PerformanceAnalytics that
compatible with table.Performance
}
\details{
use \code{table.Performance.pool} to check available
metrics. recoded SharpeRatio
}
\examples{
table.Performance.pool()
}
\author{
Kirk Li \email{kirkli@stat.washington.edu}
}
\seealso{
\code{\link{table.Performance}},
\code{\link{table.Performance.pool.cran}}
}
\keyword{measure}
\keyword{metrics}
\keyword{performance}
\keyword{table}

>>>>>>> a5e52e3aa047a1e03a04d2b57852d2c075519093
