% Generated by roxygen2 (4.0.0): do not edit by hand
\name{var.se}
\alias{var.se}
\title{Uncertainty measure of Variance Estimator
This function returns the standard error of the three estimator of Variance.}
\usage{
var.se(R, methods = c("normal", "non-normal", "bootstrap"))
}
\arguments{
  \item{R}{an xts, vector, matrix, data frame, timeSeries
  or zoo object of asset returns}

  \item{methods}{the estimation methods for Variance
  estimator, default is normal}
}
\description{
Uncertainty measure of Variance Estimator This function
returns the standard error of the three estimator of
Variance.
}
\examples{
data(edhec)
var.se(edhec[,1],methods="normal")
}
\author{
Douglass Martin, Kirk Li
}
\references{
TBA
}
\keyword{bootstrap}
\keyword{estimation,}
\keyword{variance}

